% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGLA.R
\name{CGL}
\alias{CGL}
\title{Cuadrado Grecolatino}
\usage{
CGL(Respuesta, Tratamiento, Filas, Columnas, LetrasGR, data)
}
\arguments{
\item{Respuesta}{(string) nombre de la variable respuesta}

\item{Tratamiento}{(string) nombre de la variable tratamiento}

\item{Filas}{(string) nombre de la variable filas}

\item{Columnas}{(string) nombre de la variable columnas}

\item{LetrasGR}{(string) nombre de la variable letras griegas}

\item{data}{(\code{data.frame}) Tabla de datos en formato largo con los datos del tratamiento, filas, columnas, letras griegas y respuesta.}
}
\value{
Devuelve una tabla en formato fata.frame con los calculos orrespondientes a una tabla Anova
}
\description{
Cuadrado Grecolatino
}
